Created basic code version that includes corrections for variations in supply voltage
To cater for any fluctuations in the supply voltage (also used to feed the AVCC pin on the ATMega328), we get funky with the ADC registers so that we can deal with any issues with the voltage regulator putting out a clean 5v.   

Essentially this compares the ADC reading of the known stable 1.1v internal ATmega328 reference against what the voltage is at the AVcc pin.  This gives us a correction factor created in the readVcc function.  We can then correct the signal returned from the fuel tank rheostats (as and when they are affected by a variation in the voltage being sent to the rheostats).