Create a data collection sketch
The calibration tables provided in the simple sketch are now three years old and are not adequate for our goals of accuracy.  This data collection sketch allows us to slowly fill the tank or drain the tank and take measurements by pressing a button connected to the Arduino.

The goal is to be able to build a matrix of behaviour of all three rheostats and the low fuel switch, and use this to create a highly accurate table for driving the fuel gauge.