Create AveragedCalibration.ino
New version of base code.   Averages fuel readings over 20 seconds before updating the gauge and light control.  Removes use of delay() so that the loop always runs and is driven based on counting of milliseconds and allowing for concurrent operation of the low fuel level light control and fuel level readings.